Redhat: a helper script to generate patches
Although all of my work is in git repos, there is a huge delay between
merge proposals (at LP) and releasing tarballs with the fixes. Thus, I
have to fall back to the ugly business of producing patches against the
released source tarballs.
This script does the job automatically. Regenerates all patches needed
in order to match the desired state of the submodules.