Redhat: client: require gettext, unbundle msgfmt.py
Gettext is required (spotted by Alec Leamas), as in:
http://fedoraproject.org/wiki/PackagingGuidelines#Handling_Locale_Files

msgfmt.py is a loose-in-the-net script that some projects have bundled.
For Fedora, it is packaged in python-tools (note: much irrelevant stuff
there), so it is better that we use that file instead. In order for this
to work, we have to set PYTHONPATH to the installed dir (/usr/bin).
Note that this file is only used during packaging. It is not shipped.